Add pause + keys, fix safari
I added the ability to toggle the sounds from playing to paused. I changed the sound files so they are compatible with the Safari browser. I also added event listeners to register when keys are pressed, so that the button being triggered by a keypress will be highlighted on the screen so long as that button is being held down.